Remove global parameter in VLSM.FullNode.Composite.Free
This moves the hypotheses about the clients/index type
into the section context so they are not free global
parameters any more.
Unfortunately this adds extra parameters to a few definitions
in a way that is a bit awkward for later proofs.
Perhaps they can be packaged somehow as a class instead.