Prove equivalence of step and trace has_been_sent properties
This commit shows that stepwise versions of the has_been_sent
properties are also implied by the trace-based properties in
proper_sent and proper_not_sent, giving full equivalence.

This requires slightly weaking the step property to quantify
only over protocol_transition of the
pre_loaded_with_all_messages_vlsm of the vlsm.

This commit also introduces a definition giving the properties
to ensure it's exactly the same definition in both sets of proofs.

For future work, consider revising the has_been_sent_capbility
class to use these properties, and generalizing the proofs to
work for any predicate over transition_items, not just for
has_been_sent or just for state_message_props