Make protocol_state_prop_ind usable with "induction using"
The implicit parameter {message} added from the section
was making this lemma unusable with the "induction using" tactic.
This tactic automatically finds the predicate P to use for the
induction, and also generalizes over other hypotheses that mention
the state for which we have protocol_state_prop.

This commit fixes the arguments of protocol_state_prop_ind, and changes
every place that was manually applying the lemma to instead
use "induction _ using protocol_state_prop_ind".